feat(backend): setup middleware for the validation
