+++ backend/internal/routes/users/controller.go
@@ -1,6 +1,7 @@
 package user
 
 import (
+	"db-tool/internal/core/middlewares"
 	"github.com/gin-gonic/gin"
 )
 
@@ -8,5 +9,5 @@ func Serve(r *gin.RouterGroup) {
 	user := new(User)
 	r.GET("/users", user.getAll)
 	r.GET("/users/:id", user.getByID)
-	r.POST("/users", user.create)
+	r.POST("/users", middlewares.BodyValidator[CreateUserDTO], user.create)
 }
